requirements.txt
numpy
tables
Pillow
pandas